Colour theme broken into common file "colortheme_common". New example file added for CSCT theme.
LaTeX refuses to compile if optional arguments are given on more than one occasion. This means that the original model of importing the Bath theme and then redefining some choices for the CSCT theme is not viable. Instead a common base should be developed and them imported into each. This is cumbersome as it means that when a new difference is added, it must be extracted from the common file and expressed in each theme. It's probably worth looking for a more elegant way!